Fix getting stuck upon load error
Moved the timeout out of the onload event scope.  If for some reason the photo does not load the timeout is never created, and therefore, not loading another picture.  With it moved the timeout is created whether or not the photo loads.  It has prevented the module from sticking on a picture after a random time.